Convert component to tsx and add error handling
