+++ docs/BundleSize.md
@@ -1,3 +1,3 @@
-import BundleSizeComponent from "../apps/docsite/src/components/BundleSizeTable.jsx"
+import BundleSizeComponent from "../apps/docsite/src/components/BundleSizeTable.tsx"
 
 <BundleSizeComponent/>
